src/bot/requirements.txt
python-telegram-bot==21.0.1
aiofiles==24.1.0
aiosqlite==0.20.0